microPAP/requirements.txt
paho-mqtt>=1.6.1
pyproj>=3.6.0
